refactor: TokenResponse 위치 수정
- TokenManager에서 TokenResponse를 모르도록 수정
