+++ src/main/java/com/flab/eattofit/member/application/auth/AuthService.java
@@ -1,6 +1,7 @@
 package com.flab.eattofit.member.application.auth;
 
 import com.flab.eattofit.member.application.auth.dto.LoginRequest;
+import com.flab.eattofit.member.application.auth.dto.TokenResponse;
 import com.flab.eattofit.member.domain.auth.RefreshTokenRepository;
 import com.flab.eattofit.member.domain.auth.TokenManager;
 import com.flab.eattofit.member.domain.member.NicknameGenerator;
@@ -8,7 +9,6 @@ import com.flab.eattofit.member.domain.member.Member;
 import com.flab.eattofit.member.domain.member.MemberRepository;
 import com.flab.eattofit.member.infrastructure.auth.dto.OAuthProviderRequest;
 import com.flab.eattofit.member.infrastructure.auth.dto.OAuthUserResponse;
-import com.flab.eattofit.member.infrastructure.auth.dto.TokenResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -29,10 +29,11 @@ public class AuthService {
         Member member = memberRepository.findByEmail(response.email())
                 .orElseGet(() -> registerOAuthMember(response.email(), response.name()));
 
-        TokenResponse tokens = tokenManager.getUserToken(member.getId());
-        refreshTokenRepository.save(member.getId(), tokens.refreshToken());
+        String accessToken = tokenManager.generateAccessToken(member.getId());
+        String refreshToken = tokenManager.generateRefreshToken();
+        refreshTokenRepository.save(member.getId(), refreshToken);
 
-        return tokens;
+        return new TokenResponse(accessToken, refreshToken);
     }
 
     private Member registerOAuthMember(final String email, final String name) {
