+++ src/main/java/com/flab/eattofit/member/application/auth/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.flab.eattofit.member.infrastructure.auth.dto;
+package com.flab.eattofit.member.application.auth.dto;
 
 public record TokenResponse(
         String accessToken,
