+++ src/main/java/com/flab/eattofit/member/domain/auth/TokenManager.java
@@ -1,9 +1,8 @@
 package com.flab.eattofit.member.domain.auth;
 
-import com.flab.eattofit.member.infrastructure.auth.dto.TokenResponse;
-
 public interface TokenManager {
 
-    TokenResponse getUserToken(Long id);
+    String generateAccessToken(Long id);
+    String generateRefreshToken();
     Long extractMemberId(String token);
 }
